Undid the view splitting
