+++ app/authorization/client/views/index.js
@@ -1,5 +1,3 @@
-import './permissionsTable.html';
-import './permissionsTable';
 import './permissions.html';
 import './permissions';
 import './permissionsRole.html';
