+++ app/authorization/client/views/permissions.js
@@ -1,3 +1,4 @@
+import { Meteor } from 'meteor/meteor';
 import _ from 'underscore';
 import s from 'underscore.string';
 import { ReactiveDict } from 'meteor/reactive-dict';
@@ -9,7 +10,7 @@ import { ChatPermissions } from '../lib/ChatPermissions';
 import { hasAllPermission } from '../hasPermission';
 import { t } from '../../../utils/client';
 import { SideNav } from '../../../ui-utils/client/lib/SideNav';
-import { CONSTANTS } from '../../lib';
+import { CONSTANTS, AuthorizationUtils } from '../../lib';
 
 import { hasAtLeastOnePermission } from '..';
 
@@ -160,4 +161,46 @@ Template.permissions.onRendered(() => {
 		SideNav.setFlex('adminFlex');
 		SideNav.openFlex();
 	});
+});
+
+Template.permissionsTable.helpers({
+	granted(roles, role) {
+		return (roles && ~roles.indexOf(role._id) && 'checked') || null;
+	},
+
+	permissionName(permission) {
+		if (permission.level === CONSTANTS.SETTINGS_LEVEL) {
+			let path = '';
+			if (permission.group) {
+				path = `${ t(permission.group) } > `;
+			}
+			if (permission.section) {
+				path = `${ path }${ t(permission.section) } > `;
+			}
+			return `${ path }${ t(permission.settingId) }`;
+		}
+
+		return t(permission._id);
+	},
+
+	permissionDescription(permission) {
+		return t(`${ permission._id }_description`);
+	},
+
+	disabled(role) {
+		return AuthorizationUtils.isRoleReadOnly(role._id);
+	},
+});
+
+Template.permissionsTable.events({
+	'click .role-permission'(e) {
+		const permissionId = e.currentTarget.getAttribute('data-permission');
+		const role = e.currentTarget.getAttribute('data-role');
+
+		const permission = permissionId && ChatPermissions.findOne(permissionId);
+
+		const action = ~permission.roles.indexOf(role) ? 'authorization:removeRoleFromPermission' : 'authorization:addPermissionToRole';
+
+		return Meteor.call(action, permissionId, role);
+	},
 });
